refactor aggregation and temp. comment out cache solution
